+++ tests/src/test/spark330/scala/org/apache/spark/sql/RapidsTestsTrait.scala
@@ -31,7 +31,7 @@
  */
 
 /*** spark-rapids-shim-json-lines
-{"spark": "331"}
+{"spark": "330"}
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql
