+++ tests/src/test/spark330/scala/org/apache/spark/sql/rapids/GpuInSubqueryExecSuite.scala
@@ -17,7 +17,7 @@
 /*** spark-rapids-shim-json-lines
 {"spark": "330"}
 {"spark": "330cdh"}
-{"spark": "331"}
+{"spark": "330"}
 {"spark": "332"}
 {"spark": "332cdh"}
 {"spark": "333"}
