+++ tests/src/test/spark330/scala/org/apache/spark/sql/rapids/suites/RapidsCastSuite.scala
@@ -15,7 +15,7 @@
  */
 
 /*** spark-rapids-shim-json-lines
-{"spark": "331"}
+{"spark": "330"}
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.suites
