+++ tests/src/test/spark330/scala/org/apache/spark/sql/rapids/suites/RapidsDataFrameAggregateSuite.scala
@@ -15,10 +15,19 @@
  */
 
 /*** spark-rapids-shim-json-lines
-{"spark": "331"}
+{"spark": "330"}
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.suites
 
-import org.apache.spark.sql.{JsonFunctionsSuite, RapidsSQLTestsTrait}
+import org.apache.spark.sql.{DataFrameAggregateSuite, RapidsSQLTestsTrait}
 
-class RapidsJsonFunctionsSuite extends JsonFunctionsSuite with RapidsSQLTestsTrait {}
+class RapidsDataFrameAggregateSuite extends DataFrameAggregateSuite with RapidsSQLTestsTrait {
+  // example to show how to replace the logic of an excluded test case in Vanilla Spark
+  testRapids("collect functions" ) {  // "collect functions" was excluded at RapidsTestSettings
+    // println("...")
+  }
+}
+
+
+
+
