+++ tests/src/test/spark330/scala/org/apache/spark/sql/rapids/suites/RapidsMathFunctionsSuite.scala
@@ -15,15 +15,11 @@
  */
 
 /*** spark-rapids-shim-json-lines
-{"spark": "331"}
+{"spark": "330"}
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.suites
 
-import org.apache.spark.sql.{RapidsSQLTestsTrait, StringFunctionsSuite}
-import org.apache.spark.sql.catalyst.expressions.ExpressionEvalHelper
+import org.apache.spark.sql.{MathFunctionsSuite, RapidsSQLTestsTrait}
 
-class RapidsStringFunctionsSuite
-  extends StringFunctionsSuite
-  with RapidsSQLTestsTrait
-  with ExpressionEvalHelper {
+class RapidsMathFunctionsSuite extends MathFunctionsSuite with RapidsSQLTestsTrait {
 }
