+++ tests/src/test/spark330/scala/org/apache/spark/sql/rapids/suites/RapidsStringFunctionsSuite.scala
@@ -15,11 +15,15 @@
  */
 
 /*** spark-rapids-shim-json-lines
-{"spark": "331"}
+{"spark": "330"}
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.suites
 
-import org.apache.spark.sql.RapidsTestsTrait
-import org.apache.spark.sql.catalyst.expressions.RegexpExpressionsSuite
+import org.apache.spark.sql.{RapidsSQLTestsTrait, StringFunctionsSuite}
+import org.apache.spark.sql.catalyst.expressions.ExpressionEvalHelper
 
-class RapidsRegexpExpressionsSuite extends RegexpExpressionsSuite with RapidsTestsTrait {}
+class RapidsStringFunctionsSuite
+  extends StringFunctionsSuite
+  with RapidsSQLTestsTrait
+  with ExpressionEvalHelper {
+}
